изменили запросы согласно тестовому заданию
было:
При создании надо было отправлять полностью заполненные структуры пациента и приобретаемых сервисов.
Согласно тестовому заданию в составе структуры создаваемого заказа надо отправлять только
id существующих пациента и услуг.
Так же добавил поле InProcessing для статуса заказа - 'в обработке'
Типы полей InProcessing и IsActive изменил с int8 на bool

иправлено - проверено

+++ internal/pkg/client/client.go
@@ -5,6 +5,6 @@ import (
 )
 
 type IRemoteStore interface {
-	GetByID(ctx context.Context, id uint64) (interface{}, error)
+	GetByID(ctx context.Context, id int) (interface{}, error)
 	GetList(ctx context.Context) (interface{}, error)
 }
